+++ src/crawler/mod.rs
@@ -3,6 +3,8 @@ pub mod spider;
 pub mod headless;
 pub mod sitemap;
 pub mod fast_spider;
+pub mod youtube;
+
 
 use headless::HeadlessBrowser;
 use std::num::NonZeroU32;